refactor: move passage-layer validation checks to polish_validation
Extract shared ValidationCheck/ValidationReport types to validation_types.py.
Move 11 passage-layer check functions (reachability, gates, routing,
prose neutrality, arc divergence, etc.) from grow_validation.py to
polish_validation.py. Add run_grow_checks() for beat-DAG only and
run_passage_checks() for passage-layer only.

Backward compatibility preserved via __getattr__ lazy re-exports in
grow_validation.py — existing imports continue to work.

Closes #999
